refactor doc mapper
move tantivy val to json to own module
remove duplicated conversion logic

move field_presence to own module
